irep.h: operand c should be fit in `uint8_t`.

+++ include/mruby/irep.h
@@ -122,7 +122,7 @@ struct mrb_insn_data {
   uint8_t insn;
   uint16_t a;
   uint16_t b;
-  uint16_t c;
+  uint8_t c;
 };
 
 struct mrb_insn_data mrb_decode_insn(const mrb_code *pc);
